img-suite/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
python-multipart>=0.0.19
Pillow>=11.0
rembg>=2.0.57
onnxruntime>=1.17
aiofiles>=24.1
